* Finally bump to C++20
* Use C++17's aligned `new` instead of `aligned_alloc()` / `_aligned_malloc()`
* Many code cleanups
* Fix `cameras.xml` XSD, make it verify with libxml2 2.12.0
* Drop obsolete workarounds for older toolchains
* Fix build on clang/i386 (on openSUSE Tumbleweed)
* Canon EOS 77D: improve black/white levels

# By Roman Lebedev (80) and others
# Via GitHub (11) and Roman Lebedev (5)
* develop: (82 commits)
  `cameras.xml`: use tabs, drop whitespaces at the end of the line
  Attempt to fix XSD validation w/ libxml2 2.12.0
  ArwDecoder::decodeRawInternal(): extract `decodeCurve()`
  ArwDecoder::decodeRawInternal(): extract `decodeTransitionalArw()`
  FileReader::readFile(): avoid performing tautological comparison
  Clang: drop hopefully-obsolete thread-safety-analysis false-positive workaround
  Cr2DecompressorImpl: when ctor-ing iterators, take things by const-ref
  NefDecoder::gammaCurve(): extract sub-statement
  NefDecoder::gammaCurve(): drop support for unused modes (i.e. other than 1)
  NefDecoder::gammaCurve(): use more early-return
  rsbench: extract `extern "C"` function out of anonymous namespace
  copyPixels(): take pointer-to-`std::byte`
  RawImageDataFloat::setWithLookUp(): take pointer-to-`std::byte`
  DngDecoder: move `getAsCFAColor()` lambda out of function, too large
  DngOpcodes::LookupOpcode: use using-declaration
  DngDecoder::parseWhiteBalance(): early return
  DngDecoder: extract `parseWhiteBalance()` from `decodeMetaDataInternal()`
  OlympusDecompressor: simplify difference of signedness check
  Use `if constexpr` where appropriate
  Use `std::is_floating_point_v` instead `std::is_floating_point<>::value`
  ...

# Conflicts:
#	.github/workflows/CI-linux.yml
#	.github/workflows/CI.yml

+++ src/librawspeed/decoders/PefDecoder.cpp
@@ -20,6 +20,7 @@
 */
 
 #include "decoders/PefDecoder.h"
+#include "adt/Casts.h"
 #include "adt/Point.h"
 #include "common/Common.h"
 #include "common/RawImage.h"
@@ -127,9 +128,9 @@ void PefDecoder::decodeMetaDataInternal(const CameraMetaData* meta) {
     const TiffEntry* wb =
         mRootIFD->getEntryRecursive(static_cast<TiffTag>(0x0201));
     if (wb->count == 4) {
-      mRaw->metadata.wbCoeffs[0] = wb->getU32(0);
-      mRaw->metadata.wbCoeffs[1] = wb->getU32(1);
-      mRaw->metadata.wbCoeffs[2] = wb->getU32(3);
+      mRaw->metadata.wbCoeffs[0] = implicit_cast<float>(wb->getU32(0));
+      mRaw->metadata.wbCoeffs[1] = implicit_cast<float>(wb->getU32(1));
+      mRaw->metadata.wbCoeffs[2] = implicit_cast<float>(wb->getU32(3));
     }
   }
 }
